Install the ReportLab fonts for barcodes
Fixes https://github.com/osiell/ansible-odoo/issues/18
As per odoo/odoo#3234, we have to install a bunch of extra fonts for the barcodes to be printed on the standard reports.
The fonts must be downloaded from the upstream reportlab website directly.
We install the fonts to `/home/{{odoo_user}}/fonts` as suggested by @sebalix.